+++ src/pages/api/educandos/index.js
@@ -1,17 +1,10 @@
-import lerPlanilhaGoogleSheetsAsync from "@/servicos/gsheets/readDataService";
-import { converterGSheetsParaEntidade } from "@/servicos/conversores/ConversorGenerico";
+import educandoRepositorio from "@/repositorios/educandosRepositorio";
 
 const ordenarPorNomeCompleto = (listaDeEducandos) =>
   listaDeEducandos.sort((a, b) => (a.nomeCompleto < b.nomeCompleto ? -1 : 1));
 
 export default async function handler(req, res) {
-  const entidade = "educandos";
-  const educandos = await lerPlanilhaGoogleSheetsAsync(`${entidade}!A:Z`);
-  const listaDeEducandos = converterGSheetsParaEntidade({
-    nomeDaEntidade: entidade,
-    dadosBrutos: educandos,
-  });
-  const listaOrdenada = ordenarPorNomeCompleto(listaDeEducandos);
-
-  res.status(200).json(listaOrdenada);
+  const todosPsEducandos = await educandoRepositorio.buscarTodosOsEducandos();
+  const todosOsEducandosOrdenados = ordenarPorNomeCompleto(todosPsEducandos);
+  res.status(200).json(todosOsEducandosOrdenados);
 }
