--- src/pages/api/insignias/[id].js
@@ -1,6 +1,0 @@
-import buscarInsigniaPorId from "@/repositorios/InsigniasRepositorio";
-
-export default async function handler(req, res) {
-  const insignia = await buscarInsigniaPorId(req.query.id);
-  res.status(200).json(insignia);
-}
