--- src/pages/api/unidades/[id].js
@@ -1,6 +1,0 @@
-import buscarUnidadePorId from "@/repositorios/UnidadesRepositorio";
-
-export default async function handler(req, res) {
-  const trilha = await buscarUnidadePorId(req.query.id);
-  res.status(200).json(trilha);
-}
